requirements.txt
apache-airflow==2.2.5
apache-airflow-providers-common-sql==1.3.2
pysqream-blue==1.0.18
